Cache
Adding a cache in as i rate limited the demo key, this is because of the way blazor components update. It will now use the cached data unless it  has been more than two hours or theres no cache data, in which case it will fetch more. Updated to use dev key now too, which is in ignored config file.